ajuste de detalhes classe Bucket e Teleop

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/drive/actuators/Bucket.java
@@ -1,6 +1,7 @@
 package org.firstinspires.ftc.teamcode.drive.actuators;
 
 import com.qualcomm.robotcore.eventloop.opmode.OpMode;
+import com.qualcomm.robotcore.hardware.HardwareMap;
 import com.qualcomm.robotcore.hardware.Servo;
 
 public class Bucket{
@@ -13,7 +14,7 @@ public class Bucket{
         left = hardwareMap.get(Servo.class,"esq");
     }
 
-    public void turn(float positionRight, float positionLeft){
+    public void turn(double positionRight, double positionLeft){
         right.setPosition(positionRight);
         left.setPosition(positionLeft);
     }
